+++ api/src/no/mnemonic/act/platform/api/request/v1/SearchFactRequest.java
@@ -49,11 +49,11 @@ public class SearchFactRequest implements ValidatingRequest {
   private Dimension dimension;
   @ApiModelProperty(value = "Include retracted Facts (default false)", example = "false")
   private Boolean includeRetracted;
-  @ApiModelProperty(value = "Only return Facts added before a specific timestamp",
+  @ApiModelProperty(value = "Only return Facts seen before a specific timestamp",
           example = "2016-09-28T21:26:22Z", dataType = "string")
   @JsonDeserialize(using = TimestampDeserializer.class)
   private Long before;
-  @ApiModelProperty(value = "Only return Facts added after a specific timestamp",
+  @ApiModelProperty(value = "Only return Facts seen after a specific timestamp",
           example = "2016-09-28T21:26:22Z", dataType = "string")
   @JsonDeserialize(using = TimestampDeserializer.class)
   private Long after;