+++ api/src/no/mnemonic/act/platform/api/request/v1/SearchObjectFactsRequest.java
@@ -49,10 +49,10 @@ public class SearchObjectFactsRequest implements ValidatingRequest {
   private Dimension dimension;
   @ApiModelProperty(value = "Include retracted Facts (default false)", example = "false")
   private Boolean includeRetracted;
-  @ApiModelProperty(value = "Only return Facts added before a specific timestamp", example = "2016-09-28T21:26:22Z", dataType = "string")
+  @ApiModelProperty(value = "Only return Facts seen before a specific timestamp", example = "2016-09-28T21:26:22Z", dataType = "string")
   @JsonDeserialize(using = TimestampDeserializer.class)
   private Long before;
-  @ApiModelProperty(value = "Only return Facts added after a specific timestamp", example = "2016-09-28T21:26:22Z", dataType = "string")
+  @ApiModelProperty(value = "Only return Facts seen after a specific timestamp", example = "2016-09-28T21:26:22Z", dataType = "string")
   @JsonDeserialize(using = TimestampDeserializer.class)
   private Long after;
   @ApiModelProperty(value = "Limit the number of returned Facts (default 25, 0 means all)", example = "25")